CORE DATA Refactoring
