+++ TheMovieApp/TheMovieApp/Resources/CellHelper.swift
@@ -2,7 +2,7 @@
 //  CellHelper.swift
 //  TheMovieApp
 //
-//  Created by Richie on 10/6/19.
+//  Created by Ricardo Montesinos on 10/6/19.
 //  Copyright © 2019 Rappi. All rights reserved.
 //
